Automatically Generate a PR with an updated openapi spec
This Workflow runs at 9am and 5pm est. It generates swagger docs for the
catena-tools-core repo, copies them into the documentation repo, and diffs the changes.
If there are any changes it will open a PR with those changes.

If a PR already exists for this branch, it will update that PR rather than making a new one.