Adding demo examples to documentation
For each of the sections, added in code snippets showing where the steps are being done in Galactic Kittens, as well as saying what scenes/scripts the changes are in.
Also took some time to set up a table for events - but I'm not sure if it's worth pursuing, or leaving it as is.